Push AUTH_FAILED to client if plugin successfully authenticated the client but connect-client script does not.
If you use radiusplugin, when the client is connecting, radiusplugin connects to radius server. If it fails to connect, client gets disconnected with AUTH_FAILED. But if radiusplugin successfully connected with radius server and client-connect script returns non-zero exit code, then client got disconnected on the server side but no disconnection message is pushed to the client, that's why client just waits for the packet for ~30 sec and tries to reconnect in a loop.

This commit fixes this issue.

Trac ticket #180 (https://community.openvpn.net/openvpn/ticket/180)